fix: moved tsx component into /components folder
Moved the AnimatedPowerPlant component inside the /components directory so that the tailwind styles are applied.
